\name{globalbinding-package}
\alias{globalbinding-package}
\alias{globalbinding}
\docType{package}
\title{\packageTitle{globalbinding}}
\description{\packageDescription{globalbinding}}
\section{Package Content}{\packageIndices{globalbinding}}
\author{\packageAuthor{globalbinding}}
\section{Maintainer}{\packageMaintainer{globalbinding}}
\keyword{package}
